Added cutom mouse bindings
Added a binding to push the floating window back into tiling, since I'd
overridden the default (mod-t). Figured it'd be best to keep all the float
related stuff on the mouse. Although I do disable the mouse sometimes, so it
may be useful to have bindings to tile floating windows in the keyboard. But we
can wait till that issue arises.