Extended bottom xmobar to remove the gap
There was a gap before between the bottom xmobar and trayer. Not sure
exactly why, perhaps an arithmetic error of some kind. Or the two were
computing percentages of a different total.

Solution was to extend the xmobar over further, then pad the text with
spaces so its not overlapped by trayer.