Workspaces displayed and cycled in lexicographical order.
Issue was that dynamically created workspaces are just stuck on the front of
the workspaces list. And since everything using workspaces displays using index
ordering by default, it is quite visible and annoying to the user (me).

Changes are to make xmobar display the workspace lexicographically and
for cycling to also be lexicographical. Only detriment is that cycling
code is slightly harder to read now, but still rather straight forward.

Prompts for workspaces (workspacePrompt) still use the broken index ordering,
but I don't see how I could fix this without modifying and compiling my own
version of XMonad (just XMonad.Prompt.Workspace in the contrib library), which
is definitely something I don't want to do. So I'm leaving it as is, since its
at worst a minor nuisance.

It is worth noting that I don't create dynamic workspaces all that often, as
any workspaces I use regularly are put in the workspaces list. So ultimately
these changes are for something that wouldn't occur very often. But I still
feel that they're worthwhile, as they are simple and handle the most egregious
aspects of the behavior.