Use dbus-launch instead of hardcoding DBUS_*_BUS_ADDRESS
Originally hardcoded environment variables DBUS_*_BUS_ADDRESS per a
reddit post / comment so as to fix xmobar. However, this turns out to
have broken launching some programs that use dbus. E.g.: simple-scan,
virt-manager. This could be worked around by using dbus-launch to launch
the programs from the terminal, but was annoying.

Debugging showed that the issue was that the DBUS_*_BUS_ADDRESS
environment variables were at fault because they were incorrect. After
digging through the manpages for `dbus-daemon` and `dbus-launch` and
some internet searching, I determined that the proper solution is to use
dbus-launch to start xmonad. That way it can setup our environment
variables for us. This allows both xmobar and applications to work.

For the record, here's an example what the error looks like with the
incorrect environment variables:

Failed to register: GDBus.Error:org.freedesktop.DBus.Error.AccessDenied:
Connection ":1.37" is not allowed to own the service
"org.gnome.SimpleScan" due to security policies in the configuration
file