Explicitly specify font to restore to default font size
The XMonad upgrade (and fixing compile errors / deprecation warnings)
changed the font size in the prompt. It became too big and not all
options fit on screen. Explicitly specifying the font resolved the
issue. I selected the font that according to the Haddock docs is the
default - so I don't know why this is necessary. But it works.